refactor dd environment
